moved tab_versions_screen.js into tab_versions.js
